+++ src/main.rs
@@ -1,11 +1,13 @@
 #[allow(unused_imports)]
 use bevy::{
     prelude::*,
-    window::{Cursor, WindowMode},
+    window::{Cursor, PrimaryWindow, WindowMode},
+    winit::WinitWindows,
 };
 use bevy_asset_loader::prelude::*;
 use bevy_inspector_egui::quick::WorldInspectorPlugin;
 use bevy_rapier2d::prelude::*;
+use winit::window::Icon;
 
 mod core;
 mod player;
@@ -66,6 +68,8 @@ fn main() {
 
     app.insert_resource(ClearColor(Color::rgb(0., 0., 0.)));
 
+    app.add_systems(Startup, set_window_icon);
+
     app.add_systems(
         OnTransition {
             from: GameState::Loading,
@@ -81,4 +85,32 @@ fn main() {
 fn setup() {
     // Good place to put window setup configs, like whether or not
     // the player has suggested the game be played fullscreen.
+}
+
+// Documented:
+// - https://bevy-cheatbook.github.io/window/icon.html
+// - https://stackoverflow.com/a/76729516/2225649
+// Open issue: https://github.com/bevyengine/bevy/issues/1031
+// TODO: Change to official approach when issue is closed and released
+fn set_window_icon(
+    main_window: Query<Entity, With<PrimaryWindow>>,
+    // we have to use `NonSend` here
+    windows: NonSend<WinitWindows>,
+) {
+    let primary = windows.get_window(main_window.single()).unwrap();
+
+    // here we use the `image` crate to load our icon data from a png file
+    // this is not a very bevy-native solution, but it will do
+    let (icon_rgba, icon_width, icon_height) = {
+        let image = image::open("assets/VerseSquircle-256.png")
+            .expect("Failed to open icon path")
+            .into_rgba8();
+        let (width, height) = image.dimensions();
+        let rgba = image.into_raw();
+        (rgba, width, height)
+    };
+
+    let icon = Icon::from_rgba(icon_rgba, icon_width, icon_height).unwrap();
+
+    primary.set_window_icon(Some(icon));
 }